refactor(template): extract secret() plumbing to a submodule
preprocessing/template.rs is promoted to a directory; the sentinel /
finalize_secrets / line-range machinery moves to template/secrets.rs
along with its dedicated tests. The TemplatePreprocessor itself, its
Preprocessor impl, the make_tracker secret-function wiring, and the
context-hash + dodot.* namespace builders stay in template/mod.rs.

Pure file move + thin re-import: no behaviour change. The new submodule
is private; nothing outside of `template/` imports SecretCallEntry,
make_secret_sentinel, or finalize_secrets directly.
